Move shield_key functions into patches
